+++ onapp2vhi/main.py
@@ -3,10 +3,12 @@ from pathlib import Path
 import click
 import onapp2vhi
 import os
+import json
 
 from onapp2vhi.utilities.config import OnApp2VHIConfig
 from onapp2vhi.utilities.template import CONFIG_TEMPLATE
 from onapp2vhi.utilities.logs.logger import setup_logger
+from onapp2vhi.inc.vinfra_wrapper import VinfraFlavor
 from onapp2vhi.utilities.config_cli import ConfigCli
 
 
@@ -18,27 +20,18 @@ def validate_flavor(ctx, param, value):
     if not value:
         return None
 
-    flavor = value.split("_")
+    onapp_flavor = {}
+    vhi = VinfraFlavor(cfg, service_user=True)
+    output = vhi.flavor_list()
+    vhi_flavors = [_flavor for _flavor in json.loads(output)]
+    for flavor in vhi_flavors:
+        if flavor["name"] == value:
+            onapp_flavor = flavor
 
-    if len(flavor) < 3:
-        raise click.BadParameter("Format must be \"name_cpus_ram\"")
+    if not onapp_flavor:
+        raise click.BadParameter(f"Flavor {value} does not exist, manually create first.")
 
-    if not flavor[1].isnumeric():
-        raise click.BadParameter("Format must be \"name_cpus_ram\", cpus must be in numeric")
-
-    if not 1 <= int(flavor[1]) <= 64:
-        raise click.BadParameter("Cpus must be in range between 1-64")
-
-    if not flavor[2].isnumeric():
-        raise click.BadParameter("Format must be \"name_cpus_ram\", ram must be in numeric")
-
-    flavor_dict = {
-        "name": value,
-        "vcpus": flavor[1],
-        "ram": flavor[2]
-    }
-
-    return flavor_dict
+    return onapp_flavor
 
 
 def search_config():