+++ CHANGELOG.md
@@ -3,10 +3,11 @@
 ### Feature
 
 - O2V-185: Implement progress bar for long running operation
+- O2V-218: refactor flavor flags
 - O2V-186: Implement cli to initialize config
 
 ### Fixes
-
+- O2V-215: Stop migrating suspended vm
 - O2V-216: Fix broken migrate multiple vm in 1 single command
 
 ## v1.0.3 (24/10/2023)